- restructured code
