+++ src/BlackWatch.Daemon/Features/PolygonApi/PolygonApiModel.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Text.Json.Serialization;
 // ReSharper disable ClassNeverInstantiated.Global
 
-namespace BlackWatch.Daemon.Features.Polygon
+namespace BlackWatch.Daemon.Features.PolygonApi
 {
     public record GroupedDailyCurrencyPricesResponse(
         string Status,
