+++ src/BlackWatch.Daemon/Features/Requests/Messari/MessariRequestRunner.cs
@@ -4,7 +4,7 @@ using BlackWatch.Daemon.RequestEngine;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace BlackWatch.Daemon.Features.Requests
+namespace BlackWatch.Daemon.Features.Requests.Messari
 {
     public class MessariRequestRunner : RequestRunner
     {
