+++ src/BlackWatch.Daemon/Features/Requests/Messari/MessariRequestRunnerOptions.cs
@@ -1,7 +1,6 @@
-using System.ComponentModel.DataAnnotations;
 using BlackWatch.Daemon.RequestEngine;
 
-namespace BlackWatch.Daemon.Features.Requests
+namespace BlackWatch.Daemon.Features.Requests.Messari
 {
     public class MessariRequestRunnerOptions : RequestRunnerOptions
     {
