+++ src/BlackWatch.Daemon/Features/Requests/Polygon/PolygonRequestRunnerOptions.cs
@@ -1,6 +1,6 @@
 using BlackWatch.Daemon.RequestEngine;
 
-namespace BlackWatch.Daemon.Features.Requests
+namespace BlackWatch.Daemon.Features.Requests.Polygon
 {
     public class PolygonRequestRunnerOptions : RequestRunnerOptions
     {
