+++ src/BlackWatch.Daemon/Program.cs
@@ -4,9 +4,11 @@ using BlackWatch.Core.Contracts;
 using BlackWatch.Core.Services;
 using BlackWatch.Daemon.Cron;
 using BlackWatch.Daemon.Features.CronActions;
-using BlackWatch.Daemon.Features.Messari;
-using BlackWatch.Daemon.Features.Polygon;
+using BlackWatch.Daemon.Features.MessariApi;
+using BlackWatch.Daemon.Features.PolygonApi;
 using BlackWatch.Daemon.Features.Requests;
+using BlackWatch.Daemon.Features.Requests.Messari;
+using BlackWatch.Daemon.Features.Requests.Polygon;
 using BlackWatch.Daemon.RequestEngine;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
