+++ src/BlackWatch.Daemon/appsettings.json
@@ -17,7 +17,8 @@
   },
   "Messari": {
     "BaseAddress": "https://data.messari.io/api/v2/",
-    "MaxRequestsPerMinute": 20
+    "MaxRequestsPerMinute": 20,
+    "QuoteLimit": 200
   },
   "Scheduling": {
     "QuoteHistoryDays": 365,
